requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
django-filter==21.1
django-widget-tweaks==1.4.12
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
djangorestframework~=3.13.1